refactor: update 컴포넌트 삭제, createForm으로 리팩토링
