+++ src/store/comment/commentSlice.ts
@@ -23,7 +23,17 @@ const commentSlice = createSlice({
       state.inputValues[action.payload.name] = action.payload.value;
     },
     editComment(state, action) {
-      state.inputValues = action.payload;
+      const target = state.comments.find(
+        comment => comment.id === action.payload,
+      );
+      if (target)
+        state.inputValues = {
+          id: target.id,
+          profile_url: target.profile_url,
+          author: target.author,
+          content: target.content,
+          createdAt: target?.createdAt,
+        };
     },
     setCurrentPage(state, action) {
       state.currentPage = action.payload;
